improvement(cron_task.py): update JP TRC image
replace the full-disk true color image with Japan true color image to decrease the upload traffic,
because one full-disk image can be as large as 100 MB at daytime. The main reason for this change is
due to the controlled upload speed to Baidu Yun, which is ~ 100 MB/s. Moreover, opening such a large
PNG file can cost quite a lot of time for a normal PC.